增补修改漏掉的

+++ src/sceneProcess.cpp
@@ -17,7 +17,7 @@ void CSceneProcess::detect(const Mat& frame, int chId)
 {	
 	struct timespec now;
 	clock_gettime(CLOCK_MONOTONIC, &now);
-	unsigned int us = now.tv_sec*1000000+now.tv_nsec/1000;
+	unsigned long long us = now.tv_sec*1000000+now.tv_nsec/1000;
 	*(unsigned long long*)frame.data = us;
 	bool bInit = false;
 	if(m_cnt[chId] == 0)
@@ -34,7 +34,7 @@ void CSceneProcess::optFlowDetect(const Mat& frame, int chId,cv::Rect &getBound)
 {	
 	struct timespec now;
 	clock_gettime(CLOCK_MONOTONIC, &now);
-	unsigned int us = now.tv_sec*1000000+now.tv_nsec/1000;
+	unsigned long long us = now.tv_sec*1000000+now.tv_nsec/1000;
 	*(unsigned long long*)frame.data = us;
 	bool bInit = true;
 	if(m_cnt[chId] == 0)
